+++ benches/pi.rs
@@ -12,7 +12,7 @@ const HIGH_PRECISION_BITS: &[usize] = &[2048, 4096, 8192];
 fn bench_pi_refinement(c: &mut Criterion) {
     if verbose() {
         for &bits in PRECISION_BITS {
-            let epsilon = UBinary::new(BigUint::from(1u32), BigInt::from(-(bits as i64)));
+            let epsilon = UBinary::new(BigUint::from(1u32), -BigInt::from(bits));
             let bounds = pi()
                 .refine_to_default(epsilon)
                 .expect("pi refinement should succeed");
@@ -24,7 +24,7 @@ fn bench_pi_refinement(c: &mut Criterion) {
     group.sample_size(10);
 
     for &bits in PRECISION_BITS {
-        let epsilon = UBinary::new(BigUint::from(1u32), BigInt::from(-(bits as i64)));
+        let epsilon = UBinary::new(BigUint::from(1u32), -BigInt::from(bits));
         group.bench_with_input(BenchmarkId::from_parameter(bits), &epsilon, |b, epsilon| {
             b.iter(|| {
                 black_box(
@@ -151,7 +151,7 @@ fn bench_sin_pi(c: &mut Criterion) {
 fn bench_pi_high_precision(c: &mut Criterion) {
     if verbose() {
         for &bits in HIGH_PRECISION_BITS {
-            let epsilon = UBinary::new(BigUint::from(1u32), BigInt::from(-(bits as i64)));
+            let epsilon = UBinary::new(BigUint::from(1u32), -BigInt::from(bits));
             let bounds = pi()
                 .refine_to_default(epsilon)
                 .expect("high precision pi should succeed");
@@ -163,7 +163,7 @@ fn bench_pi_high_precision(c: &mut Criterion) {
     group.sample_size(10);
 
     for &bits in HIGH_PRECISION_BITS {
-        let epsilon = UBinary::new(BigUint::from(1u32), BigInt::from(-(bits as i64)));
+        let epsilon = UBinary::new(BigUint::from(1u32), -BigInt::from(bits));
         group.bench_with_input(BenchmarkId::from_parameter(bits), &epsilon, |b, epsilon| {
             b.iter(|| {
                 black_box(
